refactor: analytics naming
